Refactor for combined reducers
